meta-intel-iot-security: meta-integrity: improve compiling root CA certificate into kernel
The old method did not properly trigger a rebuild when the file
changed. By adding it to SRC_URI when enabled, we get the normal
fetcher mechanism to watch out for changes and then trigger a
rebuild. Previously, the build directory was used out of fear that the
source code might be used for different kernels, sometimes with and
sometimes without the certificate. But that's not possible because
each kernel recipe has its own unique set of source files, so we can
just let the fetcher place the certificate in the source directory.

For Linux >= 4.3, the old method no longer works. Instead,
SYSTEM_TRUSTED_KEYS needs to be set. Unfortunately this has to be a
single file with (potentially) multiple keys inside. The new mechanism
in meta-integrity can only be used to set SYSTEM_TRUSTED_KEYS when
nothing else needs to add keys, otherwise the kernel needs to be
configured elsewhere.

(From meta-intel-iot-security rev: baac71ebfe20498edfeb057721161c102508fcf2)

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>