core-image-sato: SDK should contain intltool and glib-2.0 tools for NLS
The SDK (and its tests) need these tools. We happened to be finding them
from the native tools in PATH but recipe specific sysroots prevents that.

This fixes the SDK to work as originally intended and contain these tools.

We have to clear this in the eSDK case since the _append can't be overridden
from the class. Ugly, but we'll have to come back to this issue.

(From OE-core rev: edde4c52872fbd089f85b10b94db4d6ef1796be5)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>