meta-openembedded: initramfs-kexecboot-image: disable runtime dependency on run-postinsts
The runtime dependency chain leads to the install of unneeded files:

 /usr/sbin/run-postinsts
 /usr/sbin/update-rc.d

This comes from the default value of ROOTFS_BOOTSTRAP_INSTALL='run-postinsts'
To fix this, for this image we need an empty var.

While there, do the same for KERNELDEPMODDEPEND as suggested in image.bbclass.

(From meta-openembedded rev: 547c047e429408bbd8fa154038fd33a9abc099be)

Signed-off-by: Andrea Adami <andrea.adami@gmail.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>